Changes to build scripts

+++ admin/win/docker/build.sh
@@ -1,23 +1,12 @@
-#!/bin/bash -x
+#!/bin/bash
 
-
-#if [ $# -lt 1 ]; then
-#  echo "Usage: $(basename $0) directory_relative_to_home [uid]"
-#  exit
-#fi
-
-echo "before"
-cat "/etc/passwd"
+if [ $# -lt 1 ]; then
+  echo "Usage: $(basename $0) directory_relative_to_home [uid]"
+  exit
+fi
 
 useradd yonasj -u ${2:-1000}
-echo "after"
-cat "/etc/passwd"
-
-echo "Dollar Two $2 ... ${2:-1000}"
 su - yonasj << EOF
-
-  ls /home/yonasj
-
   cd /home/yonasj/$1
   rm -rf build-win32
   mkdir build-win32
@@ -28,4 +17,3 @@ su - yonasj << EOF
   make package
   ctest .
 EOF
-bg
